requirements.txt
# Core dependencies
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
requests>=2.31.0
mcp>=0.1.0
sqlparse>=0.4.4

# Database dependencies
pandas>=1.5.0,<2.0.0  # Using pandas 1.x for compatibility
psycopg2-binary>=2.9.5

# RAG tool dependencies
sentence-transformers>=2.2.2
torch>=2.2.0,<=2.2.2  # Specify a compatible PyTorch version
faiss-cpu>=1.7.4
beautifulsoup4>=4.12.2
markdown>=3.4.3
gitpython>=3.1.30
langchain>=0.0.300
numpy>=1.26.0,<2.0.0  # Using numpy 1.x for compatibility with Python 3.12

# LangGraph dependencies
langgraph>=0.0.15
google-generativeai>=0.3.0  # For Gemini integration
ollama>=0.1.0  # For Ollama integration

# Graph visualization dependencies
ipython>=8.12.0  # For IPython.display
graphviz>=0.20.1  # For graph visualization
pygraphviz>=1.10  # For advanced graph visualization
mermaid-magic>=0.1.0  # For mermaid diagram support
ipykernel==6.29.5 # For Jupyter notebook support

# MCP cli dependencies
mcp[cli]